+++ src/main/java/com/nhnacademy/book/order/dto/NonMemberOrderSaveRequestDto.java
@@ -5,7 +5,7 @@ import lombok.Getter;
 
 @AllArgsConstructor
 @Getter
-public class NonMemberOrderSaveRequestDto {
+public class NonMemberOrderSaveRequestDto{
     private String orderId;
     private String nonMemberOrderPassword;
 }
