+++ src/main/java/com/nhnacademy/book/order/service/command/OrderReturningService.java
@@ -1,0 +1,8 @@
+package com.nhnacademy.book.order.service.command;
+
+import com.nhnacademy.book.order.dto.OrderReturnRequestDto;
+
+public interface OrderReturningService {
+    String requestOrderReturn(String orderId, OrderReturnRequestDto refundRequest);
+    String completeOrderReturn(String orderId);
+}
