+++ src/main/java/com/nhnacademy/book/order/service/command/OrderWrappingService.java
@@ -1,0 +1,4 @@
+package com.nhnacademy.book.order.service.command;
+
+public interface OrderWrappingService {
+}
