+++ src/main/java/com/nhnacademy/book/orderProduct/service/OrderProductService.java
@@ -12,11 +12,4 @@ public interface OrderProductService {
     Optional<OrderProduct> findOrderProductBySellingBookId(Long sellingBookId);
     void patchStatus(Long orderProductId, OrderProductStatusPatchRequestDto patchRequest);
     void purchaseConfirmOrderProduct(Long orderProductId);
-
-    void cancelOrderProduct(Long orderProductId, Integer quantity);
-
-    //TODO: 주문상품 상태변경
-    //TODO: 주문상품 취소
-    //TODO: 주문상품 반품요청
-
 }
